feat：修复部分ui和安全问题，新增访客用户类型，通过配置自己的七牛云可进入后台发布文章-9.1

+++ My-Blog-ServeOpen/user/views/image/delete_image.py
@@ -1,25 +1,25 @@
-from django.http import HttpResponse
-# Create your views here.
 from rest_framework.response import Response
 from rest_framework.views import APIView
 from rest_framework.authentication import TokenAuthentication
-from rest_framework.permissions import AllowAny
-
+from rest_framework.permissions import IsAdminUser
 from appone.models.resource import Resource
 from luntan.tools.qiniu import delete_image
 
 
 class DeleteImageView(APIView):
-    permission_classes = [AllowAny]
+    permission_classes = [IsAdminUser]
     authentication_classes = [TokenAuthentication]
 
     def post(self, request):
         image = request.data.get('url', '')
         id = int(request.data.get('id', 0))
+        user_id = int(request.data.get('userId', 0))
+        if user_id == 0:
+            return Response({'result': "userId缺失"})
         if id:
             Resource.objects.get(id=id).delete()
-        if delete_image(image):
+        if delete_image(image, user_id):
             # 在这里可以根据需求处理删除成功后的操作，比如更新数据库等
             return Response({'result': "success"})
         else:
-            return Response({'result': "failure"})
+            return Response({'error': "七牛云删除失败"}, status=400)
